+++ actix-testing/CHANGES.md
@@ -1,5 +1,14 @@
 # Changes
 
+## [0.3.0-alpha.2] - 2019-11-xx
+
+* Re-export `test` attribute macros
+
+
+## [0.3.0-alpha.1] - 2019-11-22
+
+* Migrate to std::future
+
 ## [0.2.0] - 2019-10-14
 
 * Upgrade actix-server and actix-server-config deps